requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
motor==3.3.1
redis==5.0.1
pika==1.3.2
requests==2.31.0
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
transformers==4.35.0
torch==2.1.0
praw==7.7.1
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.3
httpx==0.25.1
aiohttp==3.9.1
